transfered aligned-buffer-functions to its own file (aligned_buffer.hpp), made sure standard-microkernel is working
